rename jest files votes to stakes
