+++ tests/constants/stakes
